fixed teacher_checker testing
